LayoutTests/inspector/styles/styles-add-blank-property-expected.txt

Tests that adding a new blank property works.

Text
Before append:
[expanded] 
element.style  { ()
font-size: 12px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;


After append:
[expanded] 
element.style  { ()
font-size: 12px;
margin-left: 3px;

======== Matched CSS Rules ========
[expanded] 
div  { (user agent stylesheet)
display: block;
